Use cloudpickle for serializing host callbacks
Add `cloudpickle` as an optional third-party dependency. This module is used
for serializing host callbacks for TPU computations on alternative JAX
backends, and currently not required or used on the standard JAX.

IFRT host callback now uses `google::protobuf::Any` instead of `std::string` as
the serialized representation for stronger type checking.

PiperOrigin-RevId: 511604542